LPS-94301 site-navigation - classes in module that are not exported should be in internal package
